+++ NEWS.md
@@ -85,3 +85,13 @@
 * Bug when specifying `sel_group` and `N` at the same time in `fc_split()`
 
 * Bug when specifying `title` in `fc_split()` with a number of splits different than two
+
+# flowchart 0.7.0
+
+* Solved `bug` when `sel_group` is used repeatedly in the same flowchart
+
+* Changed y-coordinate distribution of boxes when using `fc_stack()` with `unite = TRUE`
+
+* When performing multiple splits in a flowchart, the group label is stored in `$fc` concatenating the values of the different groups separated by '//'
+
+* Allow expressions in the label argument to produce bold or italics text, or even formulas